// src/module/task/task.enum.ts
enum StatusName {
  OPEN = "open",
  TODO = "todo",
  ON_PROGRESS = "on progress",
  DONE = "done",
}

enum TypeName {
  BUG = "bug",
  TASK = "task",
}

enum PriorityName {
  CRITICAL = "critical",
  HIGH = "high",
  MEDIUM = "medium",
  LOW = "low",
}

export {
  StatusName,
  TypeName,
  PriorityName
}